wip: doing removeCourseMutation
