+++ frontend/src/utils/apiHooks/user/index.ts
@@ -1,5 +1,6 @@
+import useRemoveCourseMutation from './useRemoveCourseMutation';
 import useUserCourses from './useUserCourses';
 import useUserDegree from './useUserDegree';
 import useUserPlanner from './useUserPlanner';
 
-export { useUserCourses, useUserDegree, useUserPlanner };
+export { useRemoveCourseMutation, useUserCourses, useUserDegree, useUserPlanner };
